Skeleton for wealth management
- Wealth.ml and Wealth.mli

- New option in main : "[2] - show wealth". Follow the directives to
reach skeleton code for buying and selling Ether

- Going to implement wealth.ml functions to add a wealth log that
maintains a simple csv with 2 values: the amount of ether the user owns
and the amount of USD they have spent purchasing this ether.

- The worth of the ether in the user's stores will be calculated in real
time, no point in keeping this in the file if it changes so often

- The functions in worth.ml do not know about any methods in
ether_scan_processing, ether_csv, etc. All this info will be passed from
main